+++ include/libhmm/training/trainers.h
@@ -11,4 +11,5 @@
 #include "libhmm/training/viterbi_trainer.h"
 #include "libhmm/training/segmental_kmeans_trainer.h"
 #include "libhmm/training/kmeans_init.h"
+#include "libhmm/training/fit_best_of_n.h"
 #include "libhmm/training/model_selection.h"